InvoiceLayoutLM Extractor/model_finetuning/requirements.txt
accelerate>=0.30.0
pillow>=10.0.0
seqeval>=1.2.2
torch>=2.2.0
transformers>=4.40.0
